+++ app/src/main/java/com/x8bit/bitwarden/ui/platform/components/coachmark/model/CoachMarkHighlightShape.kt
@@ -1,0 +1,16 @@
+package com.x8bit.bitwarden.ui.platform.components.coachmark.model
+
+/**
+ * Defines the available shapes for a coach mark highlight.
+ */
+enum class CoachMarkHighlightShape {
+    /**
+     * A square-shaped highlight.
+     */
+    SQUARE,
+
+    /**
+     * An oval-shaped highlight.
+     */
+    OVAL,
+}
